chore: rename string key-value storage module to camelCase
