+++ apps/webapp/src/script/storage/stringKeyValueStorageTypes.ts
